refactor(cli): share option builders
