+++ src/genecoder/cli/__init__.py
@@ -1,19 +1,21 @@
 """CLI submodules for GeneCoder."""
 
 from ..options import EncodingOptions, DecodingOptions
-from .encode import (
+from .options import (
     build_encoding_options,
-    run_encoding_pipeline,
-)
-from .decode import (
     build_decoding_options,
-    run_decoding_pipeline,
+    ChannelOptions,
+    build_channel_options,
 )
+from .encode import run_encoding_pipeline
+from .decode import run_decoding_pipeline
 from .cli import main
 from .channel import process_channel
 
 __all__ = [
     "EncodingOptions",
+    "ChannelOptions",
+    "build_channel_options",
     "build_encoding_options",
     "run_encoding_pipeline",
     "DecodingOptions",
