+++ src/genecoder/cli/options.py
@@ -1,0 +1,96 @@
+"""Helper functions and dataclasses for CLI argument handling."""
+
+from __future__ import annotations
+
+import argparse
+from dataclasses import dataclass
+from typing import List, Dict
+
+from genecoder.options import EncodingOptions, DecodingOptions
+
+
+@dataclass
+class ChannelOptions:
+    """Options for the ``channel`` subcommand."""
+
+    simulators: List[str]
+    constraints: Dict[str, int]
+    sub_prob: float = 0.0
+    ins_prob: float = 0.0
+    del_prob: float = 0.0
+    seed: int | None = None
+    parallel: bool = False
+    threads: int | None = None
+    processes: int | None = None
+
+
+def build_encoding_options(args: argparse.Namespace) -> EncodingOptions:
+    """Create :class:`EncodingOptions` from parsed CLI arguments."""
+
+    if not 0 <= args.gc_min <= 1 or not 0 <= args.gc_max <= 1:
+        raise ValueError("gc_min and gc_max must be between 0 and 1")
+    if args.gc_min > args.gc_max:
+        raise ValueError("gc_min cannot be greater than gc_max")
+
+    return EncodingOptions(
+        method=args.method,
+        add_parity=args.add_parity,
+        k_value=args.k_value,
+        parity_rule=args.parity_rule,
+        fec=args.fec,
+        gc_min=args.gc_min,
+        gc_max=args.gc_max,
+        max_homopolymer=args.max_homopolymer,
+        alphabet=getattr(args, "alphabet", "base4"),
+    )
+
+
+def build_decoding_options(args: argparse.Namespace) -> DecodingOptions:
+    """Create :class:`DecodingOptions` from parsed CLI arguments."""
+
+    return DecodingOptions(
+        method=args.method,
+        check_parity=args.check_parity,
+        k_value=args.k_value,
+        parity_rule=args.parity_rule,
+        alphabet=getattr(args, "alphabet", "base4"),
+    )
+
+
+def build_channel_options(args: argparse.Namespace) -> ChannelOptions:
+    """Create :class:`ChannelOptions` from parsed CLI arguments."""
+
+    from .channel import _load_config  # Local import to avoid heavy deps at import time
+
+    simulators = list(args.simulators)
+    constraints = {
+        "min_length": args.min_length,
+        "max_length": args.max_length,
+        "max_homopolymer": args.max_homopolymer,
+    }
+
+    if args.config:
+        cfg_sim, cfg_con = _load_config(args.config)
+        if cfg_sim:
+            simulators = cfg_sim
+        constraints.update(cfg_con)
+
+    prob_specified = any([args.sub_prob, args.ins_prob, args.del_prob])
+    if simulators and prob_specified:
+        raise ValueError("Probability options cannot be combined with --simulator or --config")
+    if not simulators and not prob_specified:
+        raise ValueError("At least one simulator or probability option must be specified")
+    if args.threads is not None and args.processes is not None:
+        raise ValueError("Cannot specify both --threads and --processes")
+
+    return ChannelOptions(
+        simulators=simulators,
+        constraints=constraints,
+        sub_prob=args.sub_prob,
+        ins_prob=args.ins_prob,
+        del_prob=args.del_prob,
+        seed=args.seed,
+        parallel=args.parallel,
+        threads=args.threads,
+        processes=args.processes,
+    )
