+++ tests/test_cli_helpers.py
@@ -2,12 +2,8 @@ import argparse
 import pytest
 from pathlib import Path
 from genecoder.formats import to_fasta, from_fasta
-from genecoder.cli import (
-    build_encoding_options,
-    build_decoding_options,
-    run_encoding_pipeline,
-    run_decoding_pipeline,
-)
+from genecoder.cli.options import build_encoding_options, build_decoding_options
+from genecoder.cli import run_encoding_pipeline, run_decoding_pipeline
 from genecoder.reed_solomon_codec import _HAS_REEDSOLO
 from genecoder.error_detection import PARITY_RULE_GC_EVEN_A_ODD_T
